Upgrade to SDK version 1.13.0
New function, query_counterparties, NOT covered with tests
  because original test doesn't run in NodeSE env.

Debot tests updated.

+++ src/TonClient/Modules/BocModule.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using TonSdk.Modules;
 
 /*
-* TON API version 1.12.0, boc module.
+* TON API version 1.13.0, boc module.
 * THIS FILE WAS GENERATED AUTOMATICALLY.
 */
 
@@ -206,7 +206,7 @@ namespace TonSdk.Modules
             /// Bit size of the value.
             /// </summary>
             [JsonProperty("size", NullValueHandling = NullValueHandling.Ignore)]
-            public byte Size { get; set; }
+            public uint Size { get; set; }
 
             /// <summary>
             /// e.g. `123`, `-123`. - Decimal string. e.g. `"123"`, `"-123"`.
